requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==1.11.16
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-pure-pagination==0.3.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.17.0
httplib2==0.12.0
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.11
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2016.10
pywin32-ctypes==0.2.0
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
